A pro GraphQL project structure
